pyproject.toml
[project]
name = "code-interpreter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.134.0",
    "google>=3.0.0",
    "google-genai>=1.65.0",
    "google-generativeai>=0.8.6",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
